Flask-3/wwwroot/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
line-bot-sdk==1.15.0
MarkupSafe==1.1.1
numpy==1.18.1
pandas==1.0.0
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
urllib3==1.25.8
Werkzeug==0.16.1
